+++ test/TensorOneDTest.cpp
@@ -25,8 +25,8 @@ TEST_F(TensorOneDTest, test_sum)
 {
     auto fake_math = std::make_shared<MockTensorMath>();
 
-    std::vector<float> vec_a = {1, 2, 3};
-    std::vector<float> vec_b = {6, 7};
+    std::vector<double> vec_a = {1, 2, 3};
+    std::vector<double> vec_b = {6, 7};
 
     TensorOneD tensor_a(vec_a, fake_math);
     TensorOneD tensor_b(vec_b);
@@ -40,7 +40,7 @@ TEST_F(TensorOneDTest, test_sum)
 
 TEST_F(TensorOneDTest, test_copy)
 {
-    TensorOneD one(std::vector<float>({1, 2})), two;
+    TensorOneD one(std::vector<double>({1, 2})), two;
 
     two = one;
 
@@ -58,7 +58,7 @@ TEST_F(TensorOneDTest, test_input)
 {
     TensorOneD x(3);
     x.set_dim(4);
-    std::vector<float> vals = {8, 16};
+    std::vector<double> vals = {8, 16};
     x = TensorOneD(vals);
     EXPECT_EQ(2u, x.get_dim());
     EXPECT_EQ(8, x[0]);
